Completed the system of parameters of the brain and prepared everything for the revival of Huggy

+++ Assets/Scripts/View/ScriptableObject/Enemies.cs
@@ -5,6 +5,6 @@ public class Enemies : ScriptableObject
 {
     public EnemyView Template;
     public string Name;
-    public float Damage;
-    public float Speed;       
+    public int Damage;
+    public int Speed;       
 }
